chore: move services to related directories
