+++ client/src/app/components/audit-template/audit-template-list/audit-template.component.ts
@@ -1,5 +1,5 @@
 import {Component, Inject, OnInit, Optional} from '@angular/core';
-import {AuditTemplateService} from '../../../services/audit-template.service';
+import {AuditTemplateService} from '../../../services/audits/audit-template.service';
 import {MAT_DIALOG_DATA, MatDialog, MatDialogRef} from '@angular/material/dialog';
 
 interface Template {
