+++ client/src/app/components/audit-template/audit-template-view/create-audit-template/create-audit-template.component.spec.ts
@@ -3,7 +3,7 @@ import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { FormBuilder } from '@angular/forms';
 import { RouterTestingModule } from '@angular/router/testing';
 import { CreateAuditTemplateComponent } from './create-audit-template.component';
-import { AuditTemplateService } from 'src/app/services/audit-template.service';
+import { AuditTemplateService } from 'src/app/services/audits/audit-template.service';
 
 describe('AuditTemplateComponent', () => {
   let component: CreateAuditTemplateComponent;
