+++ client/src/app/components/employees/employee-view/create-employee/create-employee.component.spec.ts
@@ -1,11 +1,11 @@
 import {ComponentFixture, TestBed} from '@angular/core/testing';
-import {AuthService} from 'src/app/services/auth.service';
-import {TokenService} from 'src/app/services/token.service';
+import {AuthService} from 'src/app/services/authentication/auth.service';
+import {TokenService} from 'src/app/services/authentication/token.service';
 import {HttpClientTestingModule} from '@angular/common/http/testing';
 import {RouterTestingModule} from '@angular/router/testing';
 import {CreateEmployeeComponent} from './create-employee.component';
 import {FormBuilder} from '@angular/forms';
-import {ManageOrganizationsService} from 'src/app/services/manage-organizations.service';
+import {ManageOrganizationsService} from 'src/app/services/organizations/manage-organizations.service';
 
 describe('SignupComponent', () => {
   let component: CreateEmployeeComponent;
