+++ client/src/app/components/home/home.component.spec.ts
@@ -1,5 +1,5 @@
 import {ComponentFixture, TestBed} from '@angular/core/testing';
-import {TokenService} from 'src/app/services/token.service';
+import {TokenService} from 'src/app/services/authentication/token.service';
 import {HttpClientTestingModule} from '@angular/common/http/testing';
 import {RouterTestingModule} from '@angular/router/testing';
 import {HomeComponent} from './home.component';
