+++ client/src/app/components/login/login.component.spec.ts
@@ -1,6 +1,6 @@
 import {ComponentFixture, TestBed} from '@angular/core/testing';
-import {AuthService} from 'src/app/services/auth.service';
-import {TokenService} from 'src/app/services/token.service';
+import {AuthService} from 'src/app/services/authentication/auth.service';
+import {TokenService} from 'src/app/services/authentication/token.service';
 import {HttpClientTestingModule} from '@angular/common/http/testing';
 import {RouterTestingModule} from '@angular/router/testing';
 import {FormBuilder} from '@angular/forms';
