+++ client/src/app/components/manage-inventory-items/manage-inventory-items.component.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { ManageInventoryItemsService } from 'src/app/services/manage-inventory-items.service';
-import { ManageAuditsService } from 'src/app/services/manage-audits.service';
+import { ManageInventoryItemsService } from 'src/app/services/inventory-items/manage-inventory-items.service';
+import { ManageAuditsService } from 'src/app/services/audits/manage-audits.service';
 import { HttpClientModule } from '@angular/common/http';
 import { RouterTestingModule } from '@angular/router/testing';
 import { ManageInventoryItemsComponent } from './manage-inventory-items.component';
