+++ client/src/app/components/review-audit/review-audit.component.spec.ts
@@ -2,8 +2,8 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { ReviewAuditComponent } from './review-audit.component';
 import { HttpClientModule } from '@angular/common/http';
 import { RouterTestingModule } from '@angular/router/testing';
-import { ManageMembersService } from 'src/app/services/manage-members.service';
-import { ManageAuditsService } from 'src/app/services/manage-audits.service';
+import { ManageMembersService } from 'src/app/services/users/manage-members.service';
+import { ManageAuditsService } from 'src/app/services/audits/manage-audits.service';
 import { FormBuilder } from '@angular/forms';
 import { MatDialogModule } from '@angular/material/dialog';
