+++ client/src/app/modules/auth/auth.module.ts
@@ -4,7 +4,7 @@ import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 import {GreetingComponent} from '../../components/greeting/greeting.component';
 import {LoginComponent} from 'src/app/components/login/login.component';
 import {CreateEmployeeComponent} from 'src/app/components/employees/employee-view/create-employee/create-employee.component';
-import {AuthService} from 'src/app/services/auth.service';
+import {AuthService} from 'src/app/services/authentication/auth.service';
 import {HttpClientModule} from '@angular/common/http';
 import {RouterModule} from '@angular/router';
 import {MatModule} from '../material/material-module';
