+++ client/src/app/services/inventory-items/manage-inventory-items.service.ts
